backend/requirements.txt
fastapi
gunicorn
uvicorn
pydantic
asyncpg
azure-identity
azure-keyvault-secrets
redis
aiohttp
apscheduler